pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-tester"
version = "3.0.0"
description = "AI-powered test case generation tool with multi-agent collaboration"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["testing", "ai", "jira", "test-automation", "multi-agent"]

dependencies = [
    "requests>=2.31.0",
    "PySide6>=6.5.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "pylint>=2.17.5",
    "mypy>=1.4.1",
    "isort>=5.12.0",
]

[project.scripts]
ai-tester = "ai_tester.main:main"
ai-tester-legacy = "ai_tester.legacy.launcher:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/ai_tester",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, may use real APIs)",
    "e2e: End-to-end tests (slowest, full system)",
    "slow: Tests that take more than 1 second",
]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | legacy
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
extend_skip = ["legacy"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = ["legacy/", "tests/"]

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0913",  # too-many-arguments
]

[tool.coverage.run]
source = ["src/ai_tester"]
omit = [
    "*/tests/*",
    "*/legacy/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
